Reestruturação do projeto. Inicialização mais rápida.
(cherry picked from commit bcaa9508e1e820340c25cc7cfe3731ddaf5e5c29)
