+++ app/src/main/java/com/example/dener/opencvdemo/State.java
@@ -1,10 +1,8 @@
 package com.example.dener.opencvdemo;
 
-/**
- * Created by dener on 09/05/2016.
- */
 public enum State {
     Initializing,
+    Initialized,
     Running,
     Paused,
     ObjectFound,
