+++ app/src/main/res/layout/show_camera.xml
@@ -23,22 +23,4 @@
         android:layout_gravity="center"
         android:visibility="gone" />
 
-    <LinearLayout
-        android:visibility="visible"
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <Switch
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/contornos"
-            android:id="@+id/switch1"
-            android:textColor="#ffffff"
-            android:paddingLeft="10dp"
-            android:paddingTop="10dp"
-            android:padding="10dp" />
-
-    </LinearLayout>
-
 </FrameLayout>
